Crate massa_signature

Signature management

Modules

error 🔒

signature_impl 🔒

Structs

PublicKeyDeserializer

Deserializer for PublicKey

PublicKeyV0

Public key used to check if a message was encoded by the corresponding PublicKey. Generated from the KeyPair using SignatureEngine

SignatureDeserializer

Serializer for Signature

Enums

KeyPair

versioned KeyPair used for signature and decryption

MassaSignatureError

Error generated by massa-signature crate

PublicKey

Public key used to check if a message was encoded by the corresponding PublicKey. Generated from the KeyPair using SignatureEngine

Signature

Signature generated from a message and a KeyPair.

Functions

verify_signature_batch

Verifies a batch of signatures