Function massa_signature::verify_signature_batch

pub fn verify_signature_batch(
    batch: &[(Hash, Signature, PublicKey)]
) -> Result<(), MassaSignatureError>

Verify a batch of signatures on a single core to gain total CPU performance. Every provided triplet (hash, signature, public_key) is verified and an error is returned if at least one of them fails.

Arguments

batch: a slice of triplets (hash, signature, public_key)

Return value

Returns Ok(()) if all signatures were successfully verified, and Err(MassaSignatureError::SignatureError(_)) if at least one of them failed.