massa_signature

Module signature_impl

Macros

KeyPair 🔒

KeyPairVariant 🔒

PublicKey 🔒

PublicKeyVariant 🔒

Signature 🔒

SignatureVariant 🔒

Structs

KeyPairV0

versioned KeyPair used for signature and decryption

PublicKeyDeserializer

Deserializer for PublicKey

PublicKeyV0

Public key used to check if a message was encoded by the corresponding PublicKey. Generated from the KeyPair using SignatureEngine

SignatureDeserializer

Serializer for Signature

SignatureV0

Signature generated from a message and a KeyPair.

Enums

KeyPair

versioned KeyPair used for signature and decryption

PublicKey

Public key used to check if a message was encoded by the corresponding PublicKey. Generated from the KeyPair using SignatureEngine

Signature

Signature generated from a message and a KeyPair.

Constants

PUBLIC_PREFIX 🔒

SECRET_PREFIX 🔒

Functions

verify_signature_batch

Verifies a batch of signatures