Struct massa_signature::signature_impl::KeyPairV0

pub struct KeyPairV0(SigningKey);

versioned KeyPair used for signature and decryption

Tuple Fields
0: SigningKey
Implementations

impl KeyPairV0

pub const SECRET_KEY_BYTES_SIZE: usize = 32usize

pub fn get_version(&self) -> u64

Return the current version keypair

pub fn get_ser_len(&self) -> usize

Return the total length after serialization

pub fn to_bytes(&self) -> Vec<u8>

Return the bytes representing the keypair (should be a reference in the future)

Example

let keypair = KeyPair::generate(0).unwrap();
let bytes = keypair.to_bytes();

impl KeyPairV0

pub fn sign(&self, hash: &Hash) -> Result<SignatureV0, MassaSignatureError>

Returns the Signature produced by signing data bytes with a KeyPair.

Example

let keypair = KeyPair::generate(0).unwrap();
let data = Hash::compute_from("Hello World!".as_bytes());
let signature = keypair.sign(&data).unwrap();

pub fn get_public_key(&self) -> PublicKeyV0

Get the public key of the keypair

Example

let keypair = KeyPair::generate(0).unwrap();
let public_key = keypair.get_public_key();

pub fn generate() -> Self

Generate a new KeyPair

Example

let keypair = KeyPair::generate(0).unwrap();
let data = Hash::compute_from("Hello World!".as_bytes());
let signature = keypair.sign(&data).unwrap();

let serialized: String = signature.to_bs58_check();

pub fn from_bytes(data: &[u8]) -> Result<Self, MassaSignatureError>

Convert a byte array of size SECRET_KEY_BYTES_SIZE to a KeyPair.

IMPORTANT: providing more bytes than needed does not result in an error.

Example

let keypair = KeyPair::generate(0).unwrap();
let bytes = keypair.to_bytes();
let keypair2 = KeyPair::from_bytes(&bytes).unwrap();