Struct massa_signature::PublicKeyV0

pub struct PublicKeyV0(VerifyingKey);

Public key used to check if a message was encoded by the corresponding PublicKey. Generated from the KeyPair using SignatureEngine

Tuple Fields
0: VerifyingKey
Implementations

impl PublicKeyV0

pub const PUBLIC_KEY_SIZE_BYTES: usize = 32usize

Size of a public key

pub fn get_ser_len(&self) -> usize

Return the total length after serialization

pub fn verify_signature( &self, hash: &Hash, signature: &SignatureV0 ) -> Result<(), MassaSignatureError>

Checks if the Signature associated with data bytes was produced with the KeyPair associated to given PublicKey

pub fn to_bytes(&self) -> Vec<u8>

Return the bytes representing the keypair (should be a reference in the future)

Example

let keypair = KeyPair::generate(0).unwrap();
let bytes = keypair.to_bytes();

pub fn from_bytes(data: &[u8]) -> Result<PublicKeyV0, MassaSignatureError>

Deserialize a PublicKey from bytes.

IMPORTANT: providing more bytes than needed does not result in an error.

Example

let keypair = KeyPair::generate(0).unwrap();

let serialized = keypair.get_public_key().to_bytes();
let deserialized: PublicKey = PublicKey::from_bytes(&serialized).unwrap();