Crate massa_signature

Signature management

Structs

PrivateKey

PrivateKey used to sign messages. Schnorr signatures require a KeyPair to be signed. The KeyPair is generated when deserializing a private key.

PublicKey

Public key used to check if a message was encoded by the corresponding PublicKey. Generated from the PrivateKey using SignatureEngine

PublicKeyDeserializer

Serializer for Signature

Signature

Signature generated from a message and a PrivateKey.

SignatureDeserializer

Serializer for Signature

Enums

MassaSignatureError

Error generated by massa-signature crate

Constants

PRIVATE_KEY_SIZE_BYTES

Size of a private key

PUBLIC_KEY_SIZE_BYTES

Size of a public key

SIGNATURE_SIZE_BYTES

Size of a signature

Functions

derive_public_key

Derives a PublicKey from a PrivateKey.

generate_random_private_key

Generate a random private key from a RNG.

sign

Returns the Signature produced by signing data bytes with a PrivateKey.

verify_signature

Checks if the Signature associated with data bytes was produced with the PrivateKey associated to given PublicKey