Module massa_signature::signature_impl

Structs

PrivateKey

Private Key used to sign messages Generated using SignatureEngine.

PublicKey

Public key used to check if a message was encoded by the corresponding PublicKey. Generated from the PrivateKey using SignatureEngine

Signature

Signature generated from a message and a privateKey.

Constants

PRIVATE_KEY_SIZE_BYTES

PUBLIC_KEY_SIZE_BYTES

SIGNATURE_SIZE_BYTES

Functions

derive_public_key

generate_random_private_key

Generate a random private key from a RNG.

sign

verify_signature